math-solver/backend/requirements.txt
fastapi
uvicorn
httpx
pillow
numpy
python-multipart